LayoutTests/imported/w3c/web-platform-tests/css/cssom/rule-restrictions-expected.txt


FAIL @page assert_equals: Transform doesn't quite apply to pages expected 1 but got 2
PASS @keyframe
